+++ POA_Website/templates/TripPage.html
@@ -18,9 +18,9 @@
         <li>Return Date</li>
     <li>{{ TripMeta[3] }}</li>
         <li>Prticpants </li>
-    <li>{{ TripMeta[6] }}</li>
-        <li>Particpants cap</li>
     <li>{{ TripMeta[7] }}</li>
+        <li>Particpants cap</li>
+    <li>{{ TripMeta[6] }}</li>
     <li><h4>Details</h4></li>
     <li>{{ Tripinfo[2]}} </li>
     <li>Cordinator Name </li>
